Update budget to the new interface
