+++ packages/budget/package.json
@@ -4,6 +4,7 @@
   "description": "Civic/Budget 2017",
   "author": "David Daniel",
   "license": "MIT",
+  "main": "src/index.js",
   "keywords": [
     "react",
     "react-router",
@@ -23,18 +24,10 @@
   "scripts": {
     "clean": "rimraf ./build && mkdir build",
     "start": "yarn start:dev",
-    "lint": "eslint --fix src tools server webpack.config.* --color",
-    "lint:watch": "npm run lint -- --watch",
     "build": "cross-env NODE_ENV=production babel-node tools/build.js",
-    "analyze": "cross-env NODE_ENV=production babel-node ./tools/analyzeBundle.js",
-    "start:dev": "cross-env NODE_ENV=development babel-node server",
-    "start:production": "npm run build && npm run start:prod:server",
-    "start:prod:server": "cross-env NODE_ENV=production babel-node server",
-    "test": "mocha --opts ./mocha.opts 'src/**/*.test.js'",
-    "test:watch": "npm run test -- --watch",
-    "coverage": "cross-env NODE_ENV=test babel-node node_modules/isparta/bin/isparta cover --root src --report html npm run test --reporter progress",
-    "test:cover:travis": "cross-env NODE_ENV=test babel-node node_modules/isparta/bin/isparta cover --root src --report lcovonly _mocha -- --require ./tools/testSetup.js src/**/*.test.js && cat ./coverage/lcov.info | node_modules/coveralls/bin/coveralls.js",
-    "show:coverage": "npm run test:cover && open coverage/index.html"
+    "start:dev": "cross-env NODE_ENV=development npx babel-node server",
+    "start:prod": "cross-env NODE_ENV=production node server",
+    "test": "cross-env BABEL_ENV=test mocha --opts ./mocha.opts 'src/**/*.test.js'"
   },
   "dependencies": {
     "@hackoregon/component-library": "^1.0.0-alpha.1c62c05f",
@@ -63,8 +56,8 @@
     "styled-components": "^1.4.2"
   },
   "devDependencies": {
-    "@hackoregon/webpacker": "^1.0.0-alpha.1c62c05f",
     "@hackoregon/civic-babel-presets": "^1.0.0-alpha.1c62c05f",
+    "@hackoregon/mock-wrapper": "^1.0.0-alpha.1c62c05f",
     "autoprefixer": "^6.7.7",
     "babel-cli": "^6.24.1",
     "chai": "^3.5.0",
@@ -78,16 +71,6 @@
     "css-loader": "^0.28.0",
     "enzyme": "^2.7.1",
     "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-babel": "^4.0.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-prefer-object-spread": "^1.1.0",
-    "eslint-plugin-react": "^6.10.0",
-    "eslint-watch": "^2.1.14",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.11.1",
-    "html-webpack-plugin": "^2.28.0",
     "isparta": "^4.0.0",
     "istanbul": "^0.4.4",
     "jest": "^20.0.4",
@@ -116,11 +99,6 @@
     "url-loader": "^0.5.7",
     "version-bump-prompt": "^3.1.0",
     "webpack": "^2.5.1",
-    "webpack-bundle-analyzer": "^2.1.1",
-    "webpack-dev-middleware": "^1.10.0",
-    "webpack-dev-server": "^2.4.1",
-    "webpack-hot-middleware": "^2.17.0",
-    "webpack-md5-hash": "^0.0.5",
-    "webpack-node-externals": "^1.5.4"
+    "webpack-blocks": "^1.0.0-rc.2"
   }
 }
