+++ packages/budget/server.js
@@ -1,0 +1,1 @@
+require('@hackoregon/dev-server')();
