--- packages/budget/src/template.ejs
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <% if (htmlWebpackPlugin.options.commitSha) { %>
-  <meta commitSha='<%= htmlWebpackPlugin.options.commitSha %>'/>
-  <% } %>
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta charset="utf-8"/>
-  <title>Hack Oregon - Budget</title>
-  <script src="https://use.fontawesome.com/e38842726f.js"></script>
-</head>
-<body style="max-width:800px; margin: auto;">
-<div id="content"></div>
-</body>
-</html>
